[Java]: on new leadership term with newer leadership term id that has same logPosition as is the case of a past leader, then simply bump the leadership term id for the canvass to be resent for newer term.

+++ aeron-cluster/src/test/java/io/aeron/cluster/ClusterFollowerTest.java
@@ -216,7 +216,7 @@ public class ClusterFollowerTest
         Thread.sleep(5000);
 
         assertThat(roleOf(leaderMemberId), is(Cluster.Role.FOLLOWER));
-        assertThat(electionCounterOf(leaderMemberId), is(NULL_VALUE));
+        assertThat(electionCounterOf(leaderMemberId), is((long)NULL_VALUE));
     }
 
     private void startNode(final int index, final boolean cleanStart)
